backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
mysql-connector-python==8.2.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-decouple==3.8
aiofiles==23.2.1
aiosmtplib==3.0.1
email-validator==2.1.0
pydantic==2.5.0